complete story 1880031
The Bagels index (also '/') and Players index pages are page-cached.
The caches expire when new bagels are players are created, updated
or deleted.  Page-caching is used because the data doesn't change
very often; we don't bake bagels often nor do we add new players often.

Also add a staging environment that can be used to measure
performance (with newrelic) without using the production environment.

Note that the staging environment uses the same database as development.